pyproject.toml
[project]
name = "er-triage-env"
version = "0.2.0"
description = "ER Triage Environment for OpenEnv Hackathon"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "fastapi",
    "uvicorn[standard]",
    "openenv-core",
    "openai",
]

[project.scripts]
server = "server.app:main"

[tool.openenv]
env_class = "server.environment.ERTriageEnvironment"